requirements.txt
numpy
pillow
opencv-python
mss
opencv-contrib-python
